+++ tests/test_stride_segmentation/test_barth_dtw.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pandas as pd
 import pytest
+from joblib import Memory
 from numpy.testing import assert_array_equal
 from pandas.testing import assert_frame_equal
 
@@ -9,6 +10,7 @@ from gaitmap.stride_segmentation import BarthDtw, create_dtw_template
 from gaitmap.utils.coordinate_conversion import convert_to_fbf
 from gaitmap.utils.datatype_helper import is_single_sensor_stride_list, is_multi_sensor_stride_list
 from tests.mixins.test_algorithm_mixin import TestAlgorithmMixin
+from tests.mixins.test_caching_mixin import TestCachingMixin
 from tests.test_stride_segmentation.test_base_dtw import (
     TestSimpleSegment,
     TestMultiDimensionalArrayInputs,
@@ -16,17 +18,24 @@ from tests.test_stride_segmentation.test_base_dtw import (
 )
 
 
-class TestMetaFunctionality(TestAlgorithmMixin):
+class MetaTestConfig:
     algorithm_class = BarthDtw
-    __test__ = True
 
     @pytest.fixture()
     def after_action_instance(self) -> BaseType:
         template = create_dtw_template(np.array([0, 1.0, 0]), sampling_rate_hz=100.0)
         dtw = self.algorithm_class(template=template, max_cost=0.5, min_match_length_s=None)
-        data = np.array([0, 1.0, 0])
+        data = np.array([0, 1.0, 0, 0])
         dtw.segment(data, sampling_rate_hz=100)
         return dtw
+
+
+class TestMetaFunctionality(MetaTestConfig, TestAlgorithmMixin):
+    __test__ = True
+
+
+class TestCachingFunctionality(MetaTestConfig, TestCachingMixin):
+    __test__ = True
 
 
 class TestRegressionOnRealData:
@@ -181,7 +190,13 @@ class TestPostProcessing:
         dtw._max_sequence_length = 1.4
 
         start_end, to_keep = dtw._postprocess_matches(
-            None, [], np.array([]), matches_start_end=example_stride_list, to_keep=to_keep, acc_cost_mat=None
+            None,
+            [],
+            np.array([]),
+            matches_start_end=example_stride_list,
+            to_keep=to_keep,
+            acc_cost_mat=None,
+            memory=Memory(None),
         )
 
         # Check that start-end is unmodified
@@ -213,7 +228,13 @@ class TestPostProcessing:
         dtw._max_sequence_length = None
 
         start_end, to_keep = dtw._postprocess_matches(
-            None, [], cost=cost, matches_start_end=example_stride_list, to_keep=to_keep, acc_cost_mat=None
+            None,
+            [],
+            cost=cost,
+            matches_start_end=example_stride_list,
+            to_keep=to_keep,
+            acc_cost_mat=None,
+            memory=Memory(None),
         )
 
         # Check that start-end is unmodified
@@ -253,7 +274,13 @@ class TestPostProcessing:
         dtw._max_sequence_length = 3.0
 
         start_end, to_keep = dtw._postprocess_matches(
-            None, [], cost=cost, matches_start_end=example_stride_list, to_keep=to_keep, acc_cost_mat=None
+            None,
+            [],
+            cost=cost,
+            matches_start_end=example_stride_list,
+            to_keep=to_keep,
+            acc_cost_mat=None,
+            memory=Memory(None),
         )
 
         # Check that start-end is unmodified
@@ -293,7 +320,13 @@ class TestPostProcessing:
         dtw._max_sequence_length = 3.0
 
         start_end, to_keep = dtw._postprocess_matches(
-            None, [], cost=cost, matches_start_end=example_stride_list, to_keep=to_keep, acc_cost_mat=None
+            None,
+            [],
+            cost=cost,
+            matches_start_end=example_stride_list,
+            to_keep=to_keep,
+            acc_cost_mat=None,
+            memory=Memory(None),
         )
 
         # Check that start-end is unmodified