+++ tests/test_stride_segmentation/test_base_dtw.py
@@ -9,31 +9,41 @@ the right thing.
 
 """
 from typing import Union, Dict
+from unittest.mock import patch
 
 import numpy as np
 import pandas as pd
 import pytest
+from numpy.testing import assert_array_equal
 
 from gaitmap.base import BaseType
 from gaitmap.stride_segmentation import BarthOriginalTemplate
-from gaitmap.stride_segmentation.base_dtw import BaseDtw
+from gaitmap.stride_segmentation.base_dtw import BaseDtw, subsequence_cost_matrix_with_constrains
 from gaitmap.stride_segmentation.dtw_templates import create_dtw_template
 from gaitmap.utils.datatype_helper import get_multi_sensor_names
 from gaitmap.utils.exceptions import ValidationError
 from tests.mixins.test_algorithm_mixin import TestAlgorithmMixin
+from tests.mixins.test_caching_mixin import TestCachingMixin
 
 
-class TestMetaFunctionality(TestAlgorithmMixin):
+class MetaTestConfig:
     algorithm_class = BaseDtw
-    __test__ = True
 
     @pytest.fixture()
     def after_action_instance(self) -> BaseType:
         template = create_dtw_template(np.array([0, 1.0, 0]), sampling_rate_hz=100.0)
-        dtw = self.algorithm_class(template=template, max_cost=0.5, min_match_length_s=None,)
-        data = np.array([0, 1.0, 0])
+        dtw = self.algorithm_class(template=template, max_cost=0.5, min_match_length_s=None)
+        data = np.array([0, 1.0, 0, 0])
         dtw.segment(data, sampling_rate_hz=100)
         return dtw
+
+
+class TestMetaFunctionality(MetaTestConfig, TestAlgorithmMixin):
+    __test__ = True
+
+
+class TestCachingFunctionality(MetaTestConfig, TestCachingMixin):
+    __test__ = True
 
 
 class DtwTestBase:
@@ -107,14 +117,27 @@ class TestIOErrors(DtwTestBase):
         assert para in str(e)
         assert str(value) in str(e)
 
-    @pytest.mark.parametrize("para", ("max_template_stretch_ms", "max_signal_stretch_ms"))
-    def test_none_constrain_is_inf(self, para):
-        template = create_dtw_template(np.array([0, 1.0, 0]), sampling_rate_hz=100.0)
-        data = np.array([*np.ones(5) * 2, 0, 1.0, 0, *np.ones(5) * 2])
-        dtw = self.init_dtw(template=template, **{para: None})
-        dtw.segment(data=data, sampling_rate_hz=100)
+    @pytest.mark.parametrize("paras", ((None, 100), (100, None)))
+    def test_none_constrain_is_inf(self, paras):
+        para_names = ("max_template_stretch_ms", "max_signal_stretch_ms")
+        func_names = ("max_subseq_steps", "max_longseq_steps")
+        none_index = list(paras).index(None)
+        paras = dict(zip(para_names, paras))
 
-        assert getattr(dtw, "_" + para.rsplit("_", 1)[0]) == np.inf
+        sampling_rate = 100
+        template = create_dtw_template(np.array([0, 1.0, 0]), sampling_rate_hz=sampling_rate)
+        data = np.array([*np.ones(5) * 2, 0, 1.0, 0, *np.ones(5) * 2])
+        dtw = self.init_dtw(template=template, **paras)
+
+        # To check if the calculation is correct we check the parameter used to call function that calculates the
+        # cost mat
+        with patch("gaitmap.stride_segmentation.base_dtw.subsequence_cost_matrix_with_constrains") as mock_func:
+            mock_func.side_effect = subsequence_cost_matrix_with_constrains
+            dtw.segment(data=data, sampling_rate_hz=sampling_rate)
+
+            args, kwargs = mock_func.call_args
+            assert kwargs[func_names[none_index]] == np.inf
+            assert kwargs[func_names[not none_index]] == paras[para_names[not none_index]] / 1000 * sampling_rate
 
     def test_constrains_without_template_sampling_rate(self):
         template = create_dtw_template(np.array([0, 1.0, 0]), sampling_rate_hz=None)
@@ -134,15 +157,28 @@ class TestIOErrors(DtwTestBase):
 
         # With resampling
         dtw = self.init_dtw(template=template, max_template_stretch_ms=1000)
-        dtw.segment(data=data, sampling_rate_hz=signal_sampling_rate)
 
-        assert dtw._max_template_stretch == 10
+        # To check if the calculation is correct we check the parameter used to call function that calculates the
+        # cost mat
+        with patch("gaitmap.stride_segmentation.base_dtw.subsequence_cost_matrix_with_constrains") as mock_func:
+            mock_func.side_effect = subsequence_cost_matrix_with_constrains
+            dtw.segment(data=data, sampling_rate_hz=signal_sampling_rate)
+
+            args, kwargs = mock_func.call_args
+            assert kwargs["max_subseq_steps"] == 10
 
         # Without resampling
         dtw = self.init_dtw(template=template, max_template_stretch_ms=1000, resample_template=False)
         dtw.segment(data=data, sampling_rate_hz=signal_sampling_rate)
 
-        assert dtw._max_template_stretch == 100
+        # To check if the calculation is correct we check the parameter used to call function that calculates the
+        # cost mat
+        with patch("gaitmap.stride_segmentation.base_dtw.subsequence_cost_matrix_with_constrains") as mock_func:
+            mock_func.side_effect = subsequence_cost_matrix_with_constrains
+            dtw.segment(data=data, sampling_rate_hz=signal_sampling_rate)
+
+            args, kwargs = mock_func.call_args
+            assert kwargs["max_subseq_steps"] == 100
 
 
 class TestSimpleSegment(DtwTestBase):