+++ CHANGELOG.md
@@ -12,9 +12,10 @@ Scientific Changes section), and this project adheres to [Semantic Versioning](h
   (https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/105)
 - New function for merging intervals that are overlapping and that are within a specific distance from each other in
   `utils.array_handling`. (https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/120)
-- Added some utils to simplify caching algorithms with `joblib`.
+- Some algorithms now support caching intermediate results with `joblib`.
   There is also a new example explaining caching in more detail.
-  (https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/126)
+  (https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/126 (mostly reverted),
+  https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/127)
 - Experimental support for ZUPT based orientation updates in the `RtsKalman` filter.
   This approach is not fully validated and should be used with care.
   If the feature is turned on, a runtime warning will indicate that as well.